+++ episuite/distributions.py
@@ -3,8 +3,6 @@ from typing import List, Optional, Union
 
 import numpy as np
 
-from episuite.durations import Durations
-
 
 class DurationDistribution(ABC):
     """Base class for a duration distribution."""
@@ -19,21 +17,16 @@ class DurationDistribution(ABC):
         raise NotImplementedError
 
 
-class DurationBootstrap(DurationDistribution):
-    """This distribution will bootstrap a specified list
-    of durations.
+class EmpiricalBootstrap(DurationDistribution):
+    """This distribution will bootstrap from an empirical
+    distribution.
 
     :param replace: if sample w/ replacement or not
     """
-    def __init__(self, durations: Union[List[int], np.ndarray],
+    def __init__(self, samples: Union[List[int], np.ndarray],
                  replace: bool = True):
-        self.durations = np.asarray(durations)
+        self.samples = np.asarray(samples)
         self.replace = replace
-
-    @classmethod
-    def from_durations(cls, duration: Durations) -> 'DurationBootstrap':
-        stay_distribution: np.ndarray = duration.get_stay_distribution()
-        return cls(stay_distribution)
 
     def sample(self, size: Optional[int] = None) -> np.ndarray:
         """Sample from the duration distribution.
@@ -41,8 +34,8 @@ class DurationBootstrap(DurationDistribution):
         :param size: amount of samples to draw.
         :returns: samples from the distribution
         """
-        size = size or len(self.durations)
-        samples = np.random.choice(self.durations,
+        size = size or len(self.samples)
+        samples = np.random.choice(self.samples,
                                    size=size,
                                    replace=self.replace)
         return samples