+++ docs/source/changelog.rst
@@ -10,8 +10,10 @@ Release v.0.2.0 `(TBD)`
 Release v.0.1.0 `(14 Mar 2021)`
 -------------------------------------------------------------------------------
     * Added more documentation;
+    * Added more tutorials;
     * Added bibliography section;
     * Added Facebook COVID-19 World Survey Data API support;
     * Added Facebook Movement Range Maps;
     * Support to Python 3.7;
+    * Added testing for MacOS/Linux/Windows;
     * Minor bug-fixes and enhancements;