+++ docs/source/changelog.rst
@@ -3,6 +3,10 @@ Changelog
 In this section you'll find information about what's new in the newer
 releases of the project.
 
+Release v.0.3.0 `(to be defined)`
+-------------------------------------------------------------------------------
+    * TODO
+
 Release v.0.2.0 `(21 Mar 2021)`
 -------------------------------------------------------------------------------
     * Added more documentation;